dma-buf/dma-fence: Allow wait_any_timeout without default_wait (v2)
dma_fence_wait_any_timeout only relies on two things to work correctly:

 1) The callback list to get called upon fence signal
 2) The SIGNALED flag to be set upon fence signal

The first if these is part of the core dma-fence API.  The second is
only mostly part of the core dma-fence API with the one issue that the
flag may not be set and dma_fence_is_signaled may never return true if
fence->ops->enable_signaling() has not been called.  It's easy enough to
work around that by always using dma_fence_is_signaled instead of
manually checking the bit and falling through to a zero-timeout wait if
none of the fences report that they are signaled.

v2:
 - Use dma_fence_is_signaled in test_signaled_any
 - Fall through to a zero-timeout wait

Cc: Christian König <christian.koenig@amd.com>
Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>