dma-buf/dma-fence: Add a mechanism for proxy fences
Proxy fences allow you to create a place-holder fence which you can
later assign to the "real" fence at which point the two look
indistinguishable to anyone other than the driver which created the
real fence.  These should be used with care as it is the responsibility
of proxy fences creator to ensure that it eventually gets assigned a
real fence so it gets signaled.

Signed-off-by: Jason Ekstrand <jason@jlekstrand.net>