selftests/memfd: Add memfd_create hugetlbfs selftest
With the addition of hugetlbfs support in memfd_create, the memfd
selftests should verify correct functionality with hugetlbfs.

Instead of writing a separate memfd hugetlbfs test, modify the
memfd_test program to take an optional argument 'hugetlbfs'.  If
the hugetlbfs argument is specified, basic memfd_create functionality
will be exercised on hugetlbfs.  If hugetlbfs is not specified, the
current functionality of the test is unchanged.

Note that many of the tests in memfd_test test file sealing operations.
hugetlbfs does not support file sealing, therefore for hugetlbfs all
sealing related tests are skipped.

In order to test on hugetlbfs, there needs to be preallocated huge pages.
A new script (run_tests) is added.  This script will first run the
existing memfd_create tests.  It will then, attempt to allocate the
required number of huge pages before running the hugetlbfs test.  At
the end of testing, it will release any huge pages allocated for testing
purposes.

Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>