spi: spidev: add exclusive bus access lock via ioctls
There are use-cases wherein one or more applications wants to communicate
with the same spi slave chip and they wish to have exclusive bus access.
May to serialize their command,response patterns.
This can be realized in two ways,
 - Identify different applications using same spi channel and implement
   synchronization mechanisms to get exclusive bus access
 - Provide some mechanism through spidev driver to get exclusive spi bus
   access like via ioctls

This patch implements the second approach which eliminates the need for
implementing complex IPC mechanisms in the applications which shares the
spi bus. This becomes even harder if we don't have access to the source
code of the application which is using the spi channel.

Signed-off-by: Vikram N <vicky773@gmail.com>