lib/string.c: check for kmalloc() failure
This is mostly to keep the number of static checker warnings down so
we can spot new bugs instead of them being drowned in noise.  This
function doesn't return normal kernel error codes but instead the return
value is used to display exactly which memory failed.  I chose -1 as
hopefully that's a helpful thing to print.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>