scsi: qedf: Fix a potential NULL pointer dereference
At the beginning of 'qedf_srr_compl()' and of 'qedf_rec_compl()', we check
if 'orig_io_req' is NULL. If this happens, a NULL pointer dereference will
occur in the error handling path.

Fix it by adding an additionnal label in the error handling path in order
to avoid this NULL pointer dereference.

Fixes: 61d8658b4a43 ("scsi: qedf: Add QLogic FastLinQ offload FCoE driver framework.")
Signed-off-by: Christophe JAILLET <christophe.jaillet@wanadoo.fr>