drm: Squelch DEBUG_ATOMIC message for pointer manipulations
These messages flood the debug logs conveying very little information
except that the same actions as performed on the last atomic modeset are
being repeated on a different pointer. That our identifier is a pointer
is an indicator that is not interesting enough to be tracked by a human ;)
An alternative would be to leave these messages compiled and give the
interesting failure/success messages a new identifier like KMS (since
they are confirmation messages similar in nature to the existing KMS
messages), or move these low-level ATOMIC operations to a new id that we
can always ignore.

A debug message that summarized the action about to be taken would be
useful. As it stands the signal:noise of a drm.debug=0x1e dmesg is
verging on the useless. We need to curb the overuse of DRM_DEBUG_ATOMIC
or stop including it in the recommended debug flags.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>