perf record: enable multiplexing scaling via -R
This patch allows perf record to request that event
timing be recorded in each sample. The timing consists
of time_enabled and time_running. These two values are
used to compute the multiplexing correction, i.e.,
how long an event was actually measured by the hardware.

To activate, the user must use:
$ perf record -a -R ....

The patch works by forcing PERF_SAMPLE_READ in raw mode (-R),
i.e., reading of the event group in each sample.

The side effect is that both time_running and time_enable are
captured + an empty four byte RAW section. This way we leverage
an existing perf record mode and do not add yet another option.

With this patch, it is possible to evaluate the total number
of occurrences of each sampling event even when multiplexing is
active.

Signed-off-by: Stephane Eranian <eranian@google.com>